refactor: move hyperopt-loss interface
