+++ freqtrade/optimize/hyperopt/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa: F401
-from freqtrade.optimize.hyperopt.hyperopt import Hyperopt, IHyperOptLoss
+from freqtrade.optimize.hyperopt.hyperopt import Hyperopt
+from freqtrade.optimize.hyperopt_loss.hyperopt_loss_interface import IHyperOptLoss
